option to update value without triggering dirty/touched
i need to update the form values without triggering dirty/touched 

i need to reset the form back to the default values (not clear it) on button click, and without this the form will be dirty and my submit button will be enabled.

what do you think about this change? i'm open to other ways to get this done too if this doesn't feel right.