`isSubmitted` and `submitCount` should only update when submit was successful
**Problem**
`isSubmitted` and `submitCount` update whenever the submit handler is triggered, regardless of whether or not the submit was successful. This should only update if the form is valid and there were no issues during the submit process.

**Solution**
Move these updates out of the `finally` block and into the `try` block, so that they only happen if there are no issues with submit.