Loosen Resolver type constraints (a bit)
The reason we need to loosen the type constraints of `Resolver` is due
to the fact that io-ts can refine a type to something else other than
`TFieldValues`. In order to get around that, we define another type
parameter which is by default instantiated to `TFieldValues` in order
to preserve backwards compatibility, but which can be overridden by
a specific implementation of `Resolver`.