[Feature Request] Make useController save to return undefined if no context or control present
Skip initialization of useController if no control or context is present. This is an attempt to solve https://github.com/react-hook-form/react-hook-form/issues/4033

The main use case would be, that users can use the hook inside of their custom component and the functionality will be silently ignored in case its not needed.