Narrow down Control type
This commit changes the `Control` type to only requires the
options in the `Control` value which are actually used by the codebase.

This commit allows compatibility between Control types which are specified
on a superset of FormFields:

```
type FormFieldsBig = { foo: string, bar: string }
type FormFieldsSmall = { foo: string } // those fields should be inside a shared component

const controlBig: Control<FormFields> = ....
const controlSubset: Control<FormFieldsSmall> = controlBig // <-- error before this commit
```

This is useful when sharing parts of a form between forms. In this case,
the shared part only knows about its own FormFields, but can now be
integrated into a larger form.

Note that the `watch` API is not compatible with larger forms (in the
way disscussed above), but as long as the usage of the `watch` API is
not required (you can exclude it when passing down the formMethods
using `Omit` / `Pick`), everything works as expected:

```
const formMethods = useFormReturn<FormFieldsBig>()
const smallFormMethods: Pick<UseFormReturn<FormFieldsSmall>, 'control' | 'register'> = formMethods
```

---

The reason why the `watch` and the `resolver` function are not subtype
compatible is the way how typescript (correctly) checks whether a
function is a subtype of another function. It does this by comparing the
parameters *contravariantly*, and the return type *covariantly*. In a
nutshell for a function to be a subtype of another function, the
parameters have to be supertypes of the original paramters and the return
type has to be a subtype of the original return type. For background, see
these two links:
- https://www.typescriptlang.org/docs/handbook/release-notes/typescript-2-6.html#strict-function-types
- https://www.stephanboyer.com/post/132/what-are-covariance-and-contravariance