Matching `ChangeHandler` type to an actual implementation
The actual implementation checks event.type and if it's not a browser event it simply consumes the whole event as a value. Right now I cannot `register` over a custom component with `onChange: (value: string)` as the typing requires my `event` to have a `target: any` property, so primitive values in `onChange` cause a typescript error.