Trigger watch callbacks in response to value changes only
Prior to 7.55.0, `watch` with a callback would trigger that callback only in response to value changes.  (This is the behavior that I expected from [the documentation](https://react-hook-form.com/docs/useform/watch): "Subscribe to field update/change," and the other forms of the `watch` API are specifically for field values.) Some code in our application depended on this behavior; for example:

```ts
useEffect(() => {
  const subscription = watch(async (formData) => {
    if (await trigger()) {
      doSomethingWithValidForm(formData);
    }
  });
  return () => subscription.unsubscribe();
}, [trigger, watch]);
```

React Hook Form 7.55 changed this behavior to trigger the watch callback on every state change, not just value changes.  Specifically, in 7f95b265d16aee8b6b34d2746f4248aebb9d4678, every occurrence of `_subjects.values.next({ values: ... })` was changed to `_subjects.state.next({ values: ... })`, which caused the watch callback to be called on every state change.  Note that, in that commit, tests in `watch.test.tsx` had to be updated to reflect the extra calls.

This means that we can get back the pre-7.55 behavior by checking whether the `values` property is present in the payload of the state change (since it was present for any actual value changes in pre-7.55).

Note that this change potentially breaks compatibility with 7.55 through 7.60 (since `watch` is no longer called on every state change), but it fixes a break in backward compatibility with pre-7.55.  I believe that this is a net improvement (especially since `watch` is deprecated starting in 7.55 and the easier-to-use `subscribe` API is now available); however, I understand if you'd prefer to not merge this.